requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.1
cron-descriptor==1.4.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-oidc-auth==0.0.5
django-oidc-provider==0.8.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.6.1
future==0.18.3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
pluggy==1.3.0
prompt-toolkit==3.0.39
pycparser==2.21
pycryptodomex==3.18.0
pyjwkest==1.4.2
pytest==7.4.2
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
schema==0.7.5
six==1.16.0
South==1.0.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
